Numerous internal, framework changes.  Fixes issue #17 Fixes issue #18.  Fixes issue #19.
