+++ .idea/modules.xml
@@ -4,6 +4,7 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/SpotifyStreamer.iml" filepath="$PROJECT_DIR$/SpotifyStreamer.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/n8droid/n8droid.iml" filepath="$PROJECT_DIR$/n8droid/n8droid.iml" />
     </modules>
   </component>
 </project>
