+++ app/build.gradle
@@ -39,4 +39,5 @@ dependencies {
     compile 'com.google.guava:guava:16.0.1'
     compile 'com.squareup:otto:1.3.5'
     compile 'com.viewpagerindicator:library:2.4.1@aar'
+    compile project(':n8droid')
 }
