+++ app/src/main/java/com/n8/spotifystreamer/events/TrackPlaybackCompleteEvent.java
@@ -1,0 +1,7 @@
+package com.n8.spotifystreamer.events;
+
+/**
+ * Created by n8 on 7/12/15.
+ */
+public class TrackPlaybackCompleteEvent {
+}
