+++ app/src/main/java/com/n8/spotifystreamer/tracks/TopTracksController.java
@@ -13,9 +13,9 @@ package com.n8.spotifystreamer.tracks;
 
 import android.support.v7.widget.LinearLayoutManager;
 
-import com.n8.spotifystreamer.BaseFragmentView;
+import com.n8.n8droid.ViewController;
 
-public interface TopTracksController extends BaseFragmentView.Controller {
+public interface TopTracksController extends ViewController {
   void onNavIconClicked();
 
   LinearLayoutManager getLinearLayoutManager();
