+++ app/src/main/res/layout/fragment_artist_search.xml
@@ -10,7 +10,7 @@
   <android.support.design.widget.CoordinatorLayout
       android:layout_width="match_parent"
       android:layout_height="match_parent"
-      android:background="@color/primary_light"
+      android:background="@color/window_background"
       tools:context="com.n8.spotifystreamer.tracks.TopTracksFragment"
       >
