+++ app/src/main/res/layout/fragment_coachmark.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context="com.n8.spotifystreamer.coachmarks.CoachmarkFragment"
-    android:background="@color/primary_light"
+    android:background="@color/window_background"
     >
     <RelativeLayout
         android:layout_width="match_parent"
