+++ app/src/main/res/layout/fragment_top_tracks.xml
@@ -9,7 +9,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:context="com.n8.spotifystreamer.tracks.TopTracksFragment"
-        android:background="@color/primary_light"
+        android:background="@color/window_background"
         >
 
         <FrameLayout
