+++ app/src/main/res/layout/recycler_view_track.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="@dimen/recycler_view_height"
     android:layout_gravity="center"
-    android:background="@color/primary_light"
+    android:background="@color/window_background"
     android:foreground="?attr/selectableItemBackground"
     android:elevation="@dimen/recycler_view_item_elevation"
     >
