+++ app/src/main/res/values/colors.xml
@@ -1,12 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="primary">#FFC107</color>
     <color name="primary_dark">#FFA000</color>
-    <color name="primary_light">#FFECB3</color>
-    <color name="accent">#795548</color>
+    <color name="primary">#FFC107</color>
 
-    <color name="accent_light">#A1887F</color>
-    <color name="accent_dark">#5D4037</color>
+    <color name="primary_foreground">#FAFAFA</color>
+
+    <color name="window_background">#FFF9C4</color>
+
+    <color name="accent_dark">#455A64</color>
+    <color name="accent">#607D8B</color>
+    <color name="accent_light">#90A4AE</color>
 
     <color name="playback_fragment_header">#DDFFFFFF</color>
 </resources>
