+++ app/src/main/res/values/styles.xml
@@ -6,7 +6,10 @@
     <style name="AppTheme.Base" parent="Theme.AppCompat.Light.NoActionBar">
         <item name="colorPrimary">@color/primary</item>
         <item name="colorPrimaryDark">@color/primary_dark</item>
+
         <item name="colorAccent">@color/accent</item>
+
+        <item name="android:windowBackground">@color/window_background</item>
 
         <item name="searchViewStyle">@style/AppTheme.SearchView</item>
     </style>
