+++ n8droid/src/androidTest/java/com/n8/n8droid/ApplicationTest.java
@@ -1,0 +1,13 @@
+package com.n8.n8droid;
+
+import android.app.Application;
+import android.test.ApplicationTestCase;
+
+/**
+ * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
+ */
+public class ApplicationTest extends ApplicationTestCase<Application> {
+  public ApplicationTest() {
+    super(Application.class);
+  }
+}
