+++ n8droid/src/main/AndroidManifest.xml
@@ -1,0 +1,10 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+          package="com.n8.n8droid">
+
+  <application android:allowBackup="true"
+               android:label="@string/app_name"
+      >
+
+  </application>
+
+</manifest>
