+++ n8droid/src/main/java/com/n8/n8droid/ViewController.java
@@ -1,0 +1,7 @@
+package com.n8.n8droid;
+
+/**
+ * Created by n8 on 7/11/15.
+ */
+public interface ViewController {
+}
