+++ n8droid/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<resources>
+  <string name="app_name">n8Droid</string>
+
+<!-- TODO: Remove or change this placeholder text -->
+  <string name="hello_blank_fragment">Hello blank fragment</string>
+</resources>
